Moved tests to separate classes
