+++ tests/Valit.Tests/ExtensionsTests/framework_tests.cs
@@ -10,50 +10,6 @@ namespace Valit.Tests.ExtensionsTests
     public class framework_tests
     {
         [Fact]
-        public void should_pass_for_int_and_uint()
-        {
-            Int16 Expected = 0;
-
-            var result = ValitRules<Model>
-                .For(_model)
-                .WithStrategy(ValitRulesStrategies.Complete)
-                .Ensure(m => m.Int16, _ => _
-                    .IsGreaterThan(Int16.MinValue)
-                    .IsLessThan(Int16.MaxValue)
-                    .IsEqualTo(Expected)
-                    .WithMessage("Int16"))
-                .Ensure(m => m.Int32, _ => _
-                    .IsGreaterThan(Int32.MinValue)
-                    .IsLessThan(Int32.MaxValue)
-                    .IsEqualTo(0)
-                    .WithMessage("Int32"))
-                .Ensure(m => m.Int64, _ => _
-                    .IsGreaterThan(Int64.MinValue)
-                    .IsLessThan(Int64.MaxValue)
-                    .IsEqualTo(0)
-                    .WithMessage("Int64"))
-                .Ensure(m => m.UInt16, _ => _
-                    .IsGreaterThan(UInt16.MinValue)
-                    .IsLessThan(UInt16.MaxValue)
-                    .IsEqualTo<Model, UInt16>(1)
-                    .WithMessage("UInt16"))
-                .Ensure(m => m.UInt32, _ => _
-                    .IsGreaterThan(UInt32.MinValue)
-                    .IsLessThan(UInt32.MaxValue)
-                    .IsEqualTo<Model, UInt32>(1)
-                    .WithMessage("UInt32"))
-                .Ensure(m => m.UInt64, _ => _
-                    .IsGreaterThan(UInt64.MinValue)
-                    .IsLessThan(UInt64.MaxValue)
-                    .IsEqualTo<Model, UInt64>(1)
-                    .WithMessage("UInt64"))
-                .Validate();
-
-            Assert.Equal(true, result.Succeeded);
-            Assert.Equal("", string.Join(", ", result.Errors ?? new string[1]));
-        }
-
-        [Fact]
         public void should_not_add_messages()
         {
             var result = ValitRules<Model>
@@ -94,7 +50,8 @@ namespace Valit.Tests.ExtensionsTests
                 .Validate();
 
             Assert.Equal(false, result1.Succeeded);
-            Assert.Equal("Should add message 1", string.Join(", ", result1.Errors ?? new string[1]));
+            Assert.Equal("Should add message 1", 
+                string.Join(", ", result1.Errors ?? new string[1]));
         }
 
         [Fact]
@@ -116,7 +73,8 @@ namespace Valit.Tests.ExtensionsTests
                 .Validate();
 
             Assert.Equal(false, result.Succeeded);
-            Assert.Equal("Should add message 2", string.Join(", ", result.Errors ?? new string[1]));
+            Assert.Equal("Should add message 2", 
+                string.Join(", ", result.Errors ?? new string[1]));
         }
 
         [Fact]
@@ -138,7 +96,8 @@ namespace Valit.Tests.ExtensionsTests
                 .Validate();
 
             Assert.Equal(false, result.Succeeded);
-            Assert.Equal("Should add message 1, Should add message 2", string.Join(", ", result.Errors ?? new string[1]));
+            Assert.Equal("Should add message 1, Should add message 2", 
+                string.Join(", ", result.Errors ?? new string[1]));
         }
 
         [Fact]
@@ -160,7 +119,8 @@ namespace Valit.Tests.ExtensionsTests
                 .Validate();
 
             Assert.Equal(false, result.Succeeded);
-            Assert.Equal("Should add message 3", string.Join(", ", result.Errors ?? new string[1]));
+            Assert.Equal("Should add message 3", 
+                string.Join(", ", result.Errors ?? new string[1]));
         }
 
         [Fact]
@@ -182,7 +142,8 @@ namespace Valit.Tests.ExtensionsTests
                 .Validate();
 
             Assert.Equal(false, result.Succeeded);
-            Assert.Equal("Should add message 1, Should add message 3", string.Join(", ", result.Errors ?? new string[1]));
+            Assert.Equal("Should add message 1, Should add message 3", 
+                string.Join(", ", result.Errors ?? new string[1]));
         }
 
         [Fact]
@@ -204,7 +165,8 @@ namespace Valit.Tests.ExtensionsTests
                 .Validate();
 
             Assert.Equal(false, result.Succeeded);
-            Assert.Equal("Should add message 2, Should add message 3", string.Join(", ", result.Errors ?? new string[1]));
+            Assert.Equal("Should add message 2, Should add message 3", 
+                string.Join(", ", result.Errors ?? new string[1]));
         }
 
         [Fact]
@@ -226,7 +188,8 @@ namespace Valit.Tests.ExtensionsTests
                 .Validate();
 
             Assert.Equal(false, result.Succeeded);
-            Assert.Equal("Should add message 1, Should add message 2, Should add message 3", string.Join(", ", result.Errors ?? new string[1]));
+            Assert.Equal("Should add message 1, Should add message 2, Should add message 3", 
+                string.Join(", ", result.Errors ?? new string[1]));
         }
 
         [Fact]
