+++ tests/Valit.Tests/TypeTests/int64_tests.cs
@@ -1,0 +1,40 @@
+﻿using System;
+using Xunit;
+
+namespace Valit.Tests.TypeTests
+{
+    public class int64_tests
+    {
+        [Fact]
+        public void should_pass_for_int64()
+        {
+            var result = ValitRules<object>
+                .For(0)
+                .WithStrategy(ValitRulesStrategies.Complete)
+                .Ensure(_ => Int64.Parse("0"), _ => _
+                    .IsGreaterThan(Int64.MinValue)
+                    .IsLessThan(Int64.MaxValue)
+                    .IsEqualTo(Int64.Parse("0")))
+                .Validate();
+
+            Assert.True(result.Succeeded);
+        }
+
+        [Fact]
+        public void should_not_pass_for_int64()
+        {
+            var result = ValitRules<object>
+                .For(0)
+                .WithStrategy(ValitRulesStrategies.Complete)
+                .Ensure(_ => Int64.Parse("0"), _ => _
+                    .IsGreaterThan(Int64.Parse("1"))
+                    .WithMessage("Not greater than 1")
+                    .IsLessThan(Int64.Parse("-1"))
+                    .WithMessage("Not less than -1"))
+                .Validate();
+
+            Assert.False(result.Succeeded);
+            Assert.Equal(2, result.Errors.Length);
+        }
+    }
+}
